+++ routes/image_routes.js
@@ -1,5 +1,5 @@
 const multer = require('fastify-multer');
-const ImageController = require('../controllers/image_controller');
+const ImageController = require('../controllers/image');
 
 module.exports = app => {
   const fileFilter = (req, file, cb) => {
