+++ routes/post_routes.js
@@ -1,5 +1,5 @@
-const PostController = require('../controllers/post_controller');
-const CommentController = require('../controllers/comment_controller');
+const PostController = require('../controllers/post');
+const CommentController = require('../controllers/comment');
 
 module.exports = app => {
   // Post Routes
