+++ routes/upvote_routes.js
@@ -1,4 +1,4 @@
-const UpvoteController = require('../controllers/upvote_controller');
+const UpvoteController = require('../controllers/upvote');
 
 module.exports = app => {
   app.get('/upvotes/count/:commentId', UpvoteController.getUpvoteCount);
